+++ application/modules/journal/controllers/CommentsController.php
@@ -17,7 +17,6 @@ class CommentsController extends PaperController
         $docid = (int)$this->getRequest()->getParam('docid');
         $pcid = (int)$this->getRequest()->getParam('pcid');
         $file = $this->getRequest()->getParam('file');
-        $isTmp = (boolean)$this->getRequest()->getParam('istmp');
 
         try {
             $paper = Episciences_PapersManager::get($docid, false, RVID);
@@ -39,15 +38,37 @@ class CommentsController extends PaperController
             $controllerName = 'administratepaper';
         }
 
-        $url = $this->url(['controller' => $controllerName, 'action' => 'view', 'id' => $docid ]);
+        $url = $this->url(['controller' => $controllerName, 'action' => 'view', 'id' => $paper->getLatestVersionId() ]);
         $comment = new Episciences_Comment();
 
-        if (null === $comment->find($pcid)) {
-            $message = $this->view->translate("Le commentaire demandé n’existe pas.");
-            $this->_helper->FlashMessenger->setNamespace('success')->addMessage($message);
+        try {
+            if (null === $comment->find($pcid)) {
+                $message = $this->view->translate("Le commentaire demandé n’existe pas.");
+                $this->_helper->FlashMessenger->setNamespace(Ccsd_View_Helper_Message::MSG_ERROR)->addMessage($message);
+                $this->_helper->redirector->gotoUrl($url);
+                return;
+            }
+        } catch (Zend_Json_Exception $e) {
+            trigger_error($e->getMessage());
+            return;
+        }
+
+
+        if($comment->getUid() !== Episciences_Auth::getUid()){
+            $message = $this->view->translate("Vous n'avez pas les autorisations nécessaires pour supprimer ce fichier.");
+            $this->_helper->FlashMessenger->setNamespace(Ccsd_View_Helper_Message::MSG_ERROR)->addMessage($message);
             $this->_helper->redirector->gotoUrl($url);
             return;
         }
+
+
+        if($comment->getType() === Episciences_CommentsManager::TYPE_REVISION_ANSWER_TMP_VERSION){
+            $message = $this->view->translate("Ce fichier est attaché à la version temporaire, vous ne pouvez donc pas le supprimer.");
+            $this->_helper->FlashMessenger->setNamespace(Ccsd_View_Helper_Message::MSG_ERROR)->addMessage($message);
+            $this->_helper->redirector->gotoUrl($url);
+            return;
+        }
+
 
         $isJson = Episciences_Tools::isJson($comment->getFile());
 
@@ -58,16 +79,7 @@ class CommentsController extends PaperController
             $jFiles = [];
         }
 
-        $paperId = ($paper->getPaperid()) ?: $paper->getDocid();
-
-        if ($isTmp) { // fichiers joints aux commentaires des versions temporaires
-
-            $dir = $paperId;
-            $dir .= DIRECTORY_SEPARATOR;
-            $dir .= 'tmp';
-            $dir .= DIRECTORY_SEPARATOR;
-
-        } elseif ($comment->isCopyEditingComment()) {
+        if ($comment->isCopyEditingComment()) {
             $dir = $comment->getDocid();
             $dir .= DIRECTORY_SEPARATOR;
             $dir .= Episciences_CommentsManager::COPY_EDITING_SOURCES;
@@ -82,17 +94,17 @@ class CommentsController extends PaperController
 
         $is_file = is_file($comment_path);
         if ($file && $is_file) {//note that is_file() returns false if the parent directory doesn't have +x set for you
-            if (Episciences_Auth::isLogged() && Episciences_Auth::getUid() == $comment->getUid()) {
-                $key = array_search($file, $jFiles, true);
-                if ($key !== false) {
-                    unset($jFiles[$key]);
-                }
-                !empty($jFiles) ? $comment->setFile(json_encode($jFiles)) : $comment->setFile(null);
-                unlink($comment_path);
-                $comment->save(true);
-                $message = $this->view->translate("Le fichier a bien été supprimé.");
-                $this->_helper->FlashMessenger->setNamespace('success')->addMessage($message);
+
+            $key = array_search($file, $jFiles, true);
+            if ($key !== false) {
+                unset($jFiles[$key]);
             }
+            !empty($jFiles) ? $comment->setFile(json_encode($jFiles)) : $comment->setFile(null);
+            unlink($comment_path);
+            $comment->save(true);
+            $message = $this->view->translate("Le fichier a bien été supprimé.");
+            $this->_helper->FlashMessenger->setNamespace(Ccsd_View_Helper_Message::MSG_SUCCESS)->addMessage($message);
+
         } else {
             $message = $this->view->translate("Impossible de supprimer le fichier : élément introuvable.");
             $this->_helper->FlashMessenger->setNamespace(Ccsd_View_Helper_Message::MSG_ERROR)->addMessage($message);
@@ -140,8 +152,8 @@ class CommentsController extends PaperController
                 $form_values = $form->getValues();
                 $newComment = new Episciences_Comment();
 
-                if (isset($form_values['file_comment_author'])) { // Chargement d'un nouveau fichier
-                    $newComment->setFile($form_values['file_comment_author']);
+                if (isset($form_values[Episciences_Submit::COVER_LETTER_FILE_ELEMENT_NAME])) { // Chargement d'un nouveau fichier
+                    $newComment->setFile($form_values[Episciences_Submit::COVER_LETTER_FILE_ELEMENT_NAME]);
                     $strict = false;
                 } else {
                     $newComment->setFile($oldComment['FILE']);
@@ -150,7 +162,7 @@ class CommentsController extends PaperController
                 $newComment->setFilePath(REVIEW_FILES_PATH . $oldComment['DOCID'] . '/comments/');
                 $newComment->setType($oldComment['TYPE']);
                 $newComment->setDocid($oldComment['DOCID']);
-                $newComment->setMessage($form_values['author_comment']);
+                $newComment->setMessage($form_values[Episciences_Submit::COVER_LETTER_COMMENT_ELEMENT_NAME]);
                 $newComment->setPcid($oldComment['PCID']);
                 if (!$newComment->save($strict)) {
                     $message = $this->view->translate("Une erreur est survenue lors de l'enregistrement de votre commentaire.");
@@ -230,8 +242,8 @@ class CommentsController extends PaperController
                     $formValues = $form->getValues();
 
                     $coverLetter = [
-                        "message" => $formValues['author_comment'] ?? '',
-                        "attachedFile" => $formValues['file_comment_author'] ?? null
+                        "message" => $formValues[Episciences_Submit::COVER_LETTER_COMMENT_ELEMENT_NAME] ?? '',
+                        "attachedFile" => $formValues[Episciences_Submit::COVER_LETTER_FILE_ELEMENT_NAME] ?? null
                     ];
 
                     if (Episciences_CommentsManager::saveCoverLetter($paper, $coverLetter)) {